Organize Chinese subtitle analysis package
