+++ scinoephile/lang/zho/subtitles/analysis/__init__.py
@@ -1,14 +1,9 @@
 #  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
 #  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Code related to standard Chinese subtitles.
+"""Chinese subtitle analysis.
 
 Package hierarchy (modules may import from any above):
-* analysis
-* streams
+* result
+* cache
+* script
 """
-
-from __future__ import annotations
-
-from .analysis.result import ZhoSubtitleScriptAnalysis
-
-__all__ = ["ZhoSubtitleScriptAnalysis"]
